requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-tailwind==3.6.0
django-allauth==0.58.2
google-api-python-client==2.108.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
